+++ CHANGELOG.md
@@ -848,3 +848,4 @@
 - [2026-05-04 16:36] fix(security): prevent double tap duplicate action in security
 - [2026-05-04 16:58] chore(ui): update dependencies and cleanup unused imports in ui
 - [2026-05-04 17:21] chore(theme): update dependencies and cleanup unused imports in theme
+- [2026-05-04 17:57] perf(forms): optimize memory consumption in forms list viewport